refactor: extract DropdownMenuShell, the part of the widget that does not know what selection is
